# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CIOS Forecasting Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases/{caseId}/signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List signals for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Add a signal to a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignalRequest"
      responses:
        "201":
          description: Created signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
  /signals/{signalId}:
    put:
      operationId: updateSignal
      tags: [signals]
      summary: Update a signal
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignalRequest"
      responses:
        "200":
          description: Updated signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
    delete:
      operationId: deleteSignal
      tags: [signals]
      summary: Delete a signal
      parameters:
        - name: signalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List scenarios for all cases
      parameters:
        - name: caseId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: createScenario
      tags: [scenarios]
      summary: Create a new scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScenarioRequest"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
...
components:
  schemas:
...
    Signal:
      type: object
      properties:
        id:
          type: string
        signalId:
          type: string
        caseId:
          type: string
        candidateId:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        signalDescription:
          type: string
        signalType:
          type: string
        direction:
          type: string
        strengthScore:
          type: number
        reliabilityScore:
          type: number
        likelihoodRatio:
          type: number
        route:
          type: string
          nullable: true
        targetPopulation:
          type: string
          nullable: true
        miosFlag:
          type: string
          nullable: true
        ohosFlag:
          type: string
          nullable: true
        weightedSignalScore:
          type: number
          nullable: true
        actorAdjustedImpact:
          type: number
          nullable: true
        activeLikelihoodRatio:
          type: number
          nullable: true
        absoluteImpact:
          type: number
          nullable: true
      required:
        - id
        - signalId
        - caseId
        - signalDescription
        - signalType
        - direction
        - strengthScore
        - reliabilityScore
        - likelihoodRatio
    CreateSignalRequest:
      type: object
      properties:
        signalId:
          type: string
        signalDescription:
          type: string
        signalType:
          type: string
        direction:
          type: string
        strengthScore:
          type: number
        reliabilityScore:
          type: number
        likelihoodRatio:
          type: number
        route:
          type: string
        targetPopulation:
          type: string
        miosFlag:
          type: string
        ohosFlag:
          type: string
      required:
        - signalDescription
        - signalType
        - direction
        - strengthScore
        - reliabilityScore
        - likelihoodRatio
...
    Scenario:
      type: object
      properties:
        id:
          type: string
        scenarioId:
          type: string
        caseId:
          type: string
        hypotheticalSignal:
          type: string
        direction:
          type: string
        estimatedImpact:
          type: number
        newProbability:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - scenarioId
        - caseId
        - hypotheticalSignal
        - direction
        - estimatedImpact
    CreateScenarioRequest:
      type: object
      properties:
        caseId:
          type: string
        hypotheticalSignal:
          type: string
        direction:
          type: string
        estimatedImpact:
          type: number
        notes:
          type: string
      required:
        - caseId
        - hypotheticalSignal
        - direction
        - estimatedImpact
...